% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gse_workflow.R
\name{process_gse}
\alias{process_gse}
\title{Download GSE file with annotations and extract expression data}
\usage{
process_gse(geo_id, destdir = NULL, outputdir = NULL, control = NULL,
  treatment = NULL)
}
\arguments{
\item{geo_id}{character}

\item{destdir}{cache directory}

\item{outputdir}{place to store output files}

\item{control}{vector of sample ids}

\item{treatment}{vector of sample ids}
}
\value{
vector of output files
}
\description{
Download GSE file with annotations and extract expression data
}

